updated for code climate analysis and added drop-frame attribution
