+++ vtc/_timecode_sections.py
@@ -1,0 +1,12 @@
+from typing import NamedTuple
+
+
+class TimecodeSections(NamedTuple):
+    """
+    TimecodeSections contains the sections of a parsed timecode value as ints.
+    """
+
+    hours: int
+    minutes: int
+    seconds: int
+    frames: int
